refactor(mempool-service): split service and coherence into per-section modules
Move-only split of service.rs and coherence.rs along their existing
`// ---- ... ----` section dividers into submodules, per the "sections are
modules" convention. No behavior, signature, or public-API change.
